Completadas las funciones para leer los archivos, y buscar un hechizo en hechizos.txt solo ingresando su nombre ademas completada la primera funcion del menu de analisis

+++ src/Taller3/CalcularPuntaje.java
@@ -1,5 +1,5 @@
 package Taller3;
 
 public interface CalcularPuntaje {
-	double calcularPuntaje();
+	int calcularPuntaje();
 }
